+++ tencentcloud/bda/v20200324/client.go
@@ -370,6 +370,31 @@ func (c *Client) SearchTrace(request *SearchTraceRequest) (response *SearchTrace
     return
 }
 
+func NewSegmentCustomizedPortraitPicRequest() (request *SegmentCustomizedPortraitPicRequest) {
+    request = &SegmentCustomizedPortraitPicRequest{
+        BaseRequest: &tchttp.BaseRequest{},
+    }
+    request.Init().WithApiInfo("bda", APIVersion, "SegmentCustomizedPortraitPic")
+    return
+}
+
+func NewSegmentCustomizedPortraitPicResponse() (response *SegmentCustomizedPortraitPicResponse) {
+    response = &SegmentCustomizedPortraitPicResponse{
+        BaseResponse: &tchttp.BaseResponse{},
+    }
+    return
+}
+
+// 在前后景分割的基础上优化多分类分割，支持对头发、五官等的分割，既作为换发型、挂件等底层技术，也可用于扣人头、扣人脸等玩法
+func (c *Client) SegmentCustomizedPortraitPic(request *SegmentCustomizedPortraitPicRequest) (response *SegmentCustomizedPortraitPicResponse, err error) {
+    if request == nil {
+        request = NewSegmentCustomizedPortraitPicRequest()
+    }
+    response = NewSegmentCustomizedPortraitPicResponse()
+    err = c.Send(request, response)
+    return
+}
+
 func NewSegmentPortraitPicRequest() (request *SegmentPortraitPicRequest) {
     request = &SegmentPortraitPicRequest{
         BaseRequest: &tchttp.BaseRequest{},