+++ CHANGELOG.md
@@ -1,52 +1,99 @@
 # 本版本更新包含以下内容：
 
-## 容器镜像服务(tcr) 版本：2019-09-24
+## 人体分析(bda) 版本：2020-03-24
 
-### 第 13 次发布
+### 第 4 次发布
 
-发布时间：2020-07-29 08:21:59
+发布时间：2020-07-30 08:01:28
 
 本次发布包含了以下内容：
 
 改善已有的文档。
 
-修改接口：
+新增接口：
 
-* [CreateInstance](https://cloud.tencent.com/document/api/1141/41572)
-
-	* 新增入参：TagSpecification
-
+* [SegmentCustomizedPortraitPic](https://cloud.tencent.com/document/api/1208/47022)
 
 新增数据结构：
 
-* [Tag](https://cloud.tencent.com/document/api/1141/41603#Tag)
-* [TagSpecification](https://cloud.tencent.com/document/api/1141/41603#TagSpecification)
-
-修改数据结构：
-
-* [Registry](https://cloud.tencent.com/document/api/1141/41603#Registry)
-
-	* 新增成员：TagSpecification
+* [SegmentationOptions](https://cloud.tencent.com/document/api/1208/42985#SegmentationOptions)
 
 
 
+## 计费相关(billing) 版本：2018-07-09
 
-## 私有网络(vpc) 版本：2017-03-12
+### 第 20 次发布
 
-### 第 54 次发布
-
-发布时间：2020-07-29 08:19:25
+发布时间：2020-07-30 08:04:30
 
 本次发布包含了以下内容：
 
 改善已有的文档。
 
-修改接口：
+修改数据结构：
 
-* [CreateNetDetect](https://cloud.tencent.com/document/api/215/38699)
+* [BillTransactionInfo](https://cloud.tencent.com/document/api/555/19183#BillTransactionInfo)
 
-	* <font color="#dd0000">**修改入参**：</font>NextHopType, NextHopDestination
+	* 新增成员：PayChannel, DeductMode
 
+
+
+
+## 内容安全(cms) 版本：2019-03-21
+
+### 第 15 次发布
+
+发布时间：2020-07-30 08:07:34
+
+本次发布包含了以下内容：
+
+改善已有的文档。
+
+修改数据结构：
+
+* [TextData](https://cloud.tencent.com/document/api/669/34506#TextData)
+
+	* 新增成员：DataId, Extra
+
+
+
+
+## 标签(tag) 版本：2018-08-13
+
+### 第 9 次发布
+
+发布时间：2020-07-30 08:15:41
+
+本次发布包含了以下内容：
+
+改善已有的文档。
+
+新增接口：
+
+* [DescribeResourcesByTagsUnion](https://cloud.tencent.com/document/api/651/47023)
+
+
+
+## 文本内容检测(tms) 版本：2020-07-13
+
+### 第 1 次发布
+
+发布时间：2020-07-29 11:02:07
+
+本次发布包含了以下内容：
+
+改善已有的文档。
+
+新增接口：
+
+* [TextModeration](https://cloud.tencent.com/document/api/1124/46976)
+
+新增数据结构：
+
+* [DetailResults](https://cloud.tencent.com/document/api/1124/46977#DetailResults)
+* [Device](https://cloud.tencent.com/document/api/1124/46977#Device)
+* [RiskDetails](https://cloud.tencent.com/document/api/1124/46977#RiskDetails)
+* [User](https://cloud.tencent.com/document/api/1124/46977#User)
 
 
 
@@ -2195,6 +2242,22 @@
 
 ## 人体分析(bda) 版本：2020-03-24
 
+### 第 4 次发布
+
+发布时间：2020-07-30 08:01:28
+
+本次发布包含了以下内容：
+
+改善已有的文档。
+
+新增接口：
+
+* [[SegmentCustomizedPortraitPic](https://cloud.tencent.com/document/api/1208/47022)](https://cloud.tencent.com/document/api/#/#)
+
+新增数据结构：
+
+* [[SegmentationOptions](https://cloud.tencent.com/document/api/1208/42985#SegmentationOptions)](https://cloud.tencent.com/document/api/1208/42985#[SegmentationOptions](https://cloud.tencent.com/document/api/1208/42985#SegmentationOptions))
+
 ### 第 3 次发布
 
 发布时间：2020-06-30 08:01:29
@@ -2289,6 +2352,21 @@
 
 
 ## 计费相关(billing) 版本：2018-07-09
+
+### 第 20 次发布
+
+发布时间：2020-07-30 08:04:30
+
+本次发布包含了以下内容：
+
+改善已有的文档。
+
+修改数据结构：
+
+* [BillTransactionInfo](https://cloud.tencent.com/document/api/555/19183#BillTransactionInfo)
+
+	* 新增成员：PayChannel, DeductMode
+
 
 ### 第 19 次发布
 
@@ -8048,6 +8126,21 @@
 
 
 ## 内容安全(cms) 版本：2019-03-21
+
+### 第 15 次发布
+
+发布时间：2020-07-30 08:07:34
+
+本次发布包含了以下内容：
+
+改善已有的文档。
+
+修改数据结构：
+
+* [TextData](https://cloud.tencent.com/document/api/669/34506#TextData)
+
+	* 新增成员：DataId, Extra
+
 
 ### 第 14 次发布
 
@@ -23080,6 +23173,18 @@
 
 ## 标签(tag) 版本：2018-08-13
 
+### 第 9 次发布
+
+发布时间：2020-07-30 08:15:41
+
+本次发布包含了以下内容：
+
+改善已有的文档。
+
+新增接口：
+
+* [[DescribeResourcesByTagsUnion](https://cloud.tencent.com/document/api/651/47023)](https://cloud.tencent.com/document/api/#/#)
+
 ### 第 8 次发布
 
 发布时间：2020-07-14 08:17:00
@@ -24595,8 +24700,8 @@
 
 新增数据结构：
 
-* [[Tag](https://cloud.tencent.com/document/api/1141/41603#Tag)](https://cloud.tencent.com/document/api/1141/41603#[Tag](https://cloud.tencent.com/document/api/1141/41603#Tag))
-* [[TagSpecification](https://cloud.tencent.com/document/api/1141/41603#TagSpecification)](https://cloud.tencent.com/document/api/1141/41603#[TagSpecification](https://cloud.tencent.com/document/api/1141/41603#TagSpecification))
+* [Tag](https://cloud.tencent.com/document/api/1141/41603#Tag)
+* [TagSpecification](https://cloud.tencent.com/document/api/1141/41603#TagSpecification)
 
 修改数据结构：
 
@@ -26576,6 +26681,30 @@
 * [EntityRelationObject](https://cloud.tencent.com/document/api/1072/36800#EntityRelationObject)
 * [EntityRelationSubject](https://cloud.tencent.com/document/api/1072/36800#EntityRelationSubject)
 * [TripleContent](https://cloud.tencent.com/document/api/1072/36800#TripleContent)
+
+
+
+
+## 文本内容检测(tms) 版本：2020-07-13
+
+### 第 1 次发布
+
+发布时间：2020-07-29 11:02:07
+
+本次发布包含了以下内容：
+
+改善已有的文档。
+
+新增接口：
+
+* [[TextModeration](https://cloud.tencent.com/document/api/1124/46976)](https://cloud.tencent.com/document/api/#/#)
+
+新增数据结构：
+
+* [[DetailResults](https://cloud.tencent.com/document/api/1124/46977#DetailResults)](https://cloud.tencent.com/document/api/1124/46977#[DetailResults](https://cloud.tencent.com/document/api/1124/46977#DetailResults))
+* [[Device](https://cloud.tencent.com/document/api/1124/46977#Device)](https://cloud.tencent.com/document/api/1124/46977#[Device](https://cloud.tencent.com/document/api/1124/46977#Device))
+* [[RiskDetails](https://cloud.tencent.com/document/api/1124/46977#RiskDetails)](https://cloud.tencent.com/document/api/1124/46977#[RiskDetails](https://cloud.tencent.com/document/api/1124/46977#RiskDetails))
+* [[User](https://cloud.tencent.com/document/api/1124/46977#User)](https://cloud.tencent.com/document/api/1124/46977#[User](https://cloud.tencent.com/document/api/1124/46977#User))